+++ platform-api/src/internal/handler/gateway.go
@@ -44,13 +44,15 @@ var gatewayVersionPattern = regexp.MustCompile(`^([0-9]{1,6}\.[0-9]{1,6}|[0-9]{4
 // GatewayHandler handles HTTP requests for gateway operations
 type GatewayHandler struct {
 	gatewayService *service.GatewayService
+	identity       *service.IdentityService
 	slogger        *slog.Logger
 }
 
 // NewGatewayHandler creates a new gateway handler
-func NewGatewayHandler(gatewayService *service.GatewayService, slogger *slog.Logger) *GatewayHandler {
+func NewGatewayHandler(gatewayService *service.GatewayService, identity *service.IdentityService, slogger *slog.Logger) *GatewayHandler {
 	return &GatewayHandler{
 		gatewayService: gatewayService,
+		identity:       identity,
 		slogger:        slogger,
 	}
 }
@@ -105,7 +107,10 @@ func (h *GatewayHandler) CreateGateway(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	createdBy, _ := middleware.GetUserIDFromRequest(r)
+	createdBy, ok := resolveActor(w, r, h.identity, h.slogger, "register gateway")
+	if !ok {
+		return
+	}
 	gateway, err := h.gatewayService.RegisterGateway(orgId, req.Id, req.DisplayName, description, req.Endpoints,
 		isCritical, functionalityType, version, createdBy, properties)
 	if err != nil {
@@ -272,7 +277,10 @@ func (h *GatewayHandler) UpdateGateway(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	updatedBy, _ := middleware.GetUserIDFromRequest(r)
+	updatedBy, ok := resolveActor(w, r, h.identity, h.slogger, "update gateway")
+	if !ok {
+		return
+	}
 	response, err := h.gatewayService.UpdateGateway(gatewayId, orgId, updatedBy, &req)
 	if err != nil {
 		if errors.Is(err, constants.ErrGatewayNotFound) {
@@ -307,7 +315,10 @@ func (h *GatewayHandler) DeleteGateway(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	deletedBy, _ := middleware.GetUserIDFromRequest(r)
+	deletedBy, ok := resolveActor(w, r, h.identity, h.slogger, "delete gateway")
+	if !ok {
+		return
+	}
 	err := h.gatewayService.DeleteGateway(gatewayId, orgId, deletedBy)
 	if err != nil {
 		// Check for specific error types
@@ -394,7 +405,10 @@ func (h *GatewayHandler) RotateToken(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	createdBy, _ := middleware.GetUserIDFromRequest(r)
+	createdBy, ok := resolveActor(w, r, h.identity, h.slogger, "rotate gateway token")
+	if !ok {
+		return
+	}
 	response, err := h.gatewayService.RotateToken(gatewayId, orgId, createdBy)
 	if err != nil {
 		errMsg := err.Error()
@@ -446,7 +460,10 @@ func (h *GatewayHandler) RevokeToken(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	revokedBy, _ := middleware.GetUserIDFromRequest(r)
+	revokedBy, ok := resolveActor(w, r, h.identity, h.slogger, "revoke gateway token")
+	if !ok {
+		return
+	}
 	err := h.gatewayService.RevokeToken(gatewayId, tokenId, orgId, revokedBy)
 	if err != nil {
 		errMsg := err.Error()
